+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/amazon-bedrock
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies [29264a3]
+  - @ai-sdk/provider-utils@4.0.1
+  - @ai-sdk/anthropic@3.0.1
+
 ## 4.0.1
 
 ### Patch Changes