+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/rsc
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies [29264a3]
+  - @ai-sdk/provider-utils@4.0.1
+  - ai@6.0.3
+
 ## 2.0.2
 
 ### Patch Changes